Restructured internal utils package
More meaningfully named instead of generic "utils", packages should
be named in a way that informs on what they do.
